refactor: UserControllerExceptionHandler 클래스 위치 변경
사용하지 않는 파일 삭제
