--- src/main/java/techcourse/myblog/service/EmailMissException.java
@@ -1,7 +1,0 @@
-package techcourse.myblog.service;
-
-public class EmailMissException extends RuntimeException {
-    public EmailMissException(String msg) {
-        super(msg);
-    }
-}
